+++  course-work-ps-31-2013 --username dertkaes@gmail.com/OpenGLTest/OpenGLTest/view/TextWriter.h
@@ -1,5 +1,5 @@
-#include <ft2build.h>"
-#include FT_FREETYPE_H"
+#include <ft2build.h>
+#include FT_FREETYPE_H
 #include <string>
 #include <map>
 #include <set>
@@ -20,20 +20,20 @@ struct sGlyph
 	int bitmap_top;
 	int width;
 	int rows;
-	float advancex;
+	int advancex;
 };
 
 class CTextWriter
 {
 public:
 	CTextWriter();
-	void PrintText(float x, float y, std::string const& font, unsigned int size, std::string const& text, int width = 0, int height = 0);
+	void PrintText(int x, int y, std::string const& font, unsigned int size, std::string const& text, int width = 0, int height = 0);
 	int GetStringHeight(std::string const& font, unsigned int size, std::string const& text);
 	int GetStringWidth(std::string const& font, unsigned int size, std::string const& text);
 private:
 	FT_Face GetFace(std::string const& name);
 	sGlyph GetSymbol(std::string const& font, unsigned int size, char symbol);
-	void DrawBitmap(float & x, float & y, sGlyph symbol);
+	void DrawBitmap(int & x, int & y, sGlyph symbol);
 	sGlyph CreateSymbol(sSymbol s);
 	FT_Library m_ft;
 	std::map<std::string, FT_Face> m_faces;
